add action_record_function #4

+++ db/migrate/20200517124657_create_plan_allots.rb
@@ -3,7 +3,7 @@ class CreatePlanAllots < ActiveRecord::Migration[5.2]
     create_table :plan_allots do |t|
       t.integer :verb_id, null: false
       # 取り組みの配分
-      t.float :allot
+      t.integer :allot
       # 取り組み日数
       t.datetime :term
       t.timestamps
