+++ db/migrate/20200517124704_create_real_allots.rb
@@ -3,7 +3,7 @@ class CreateRealAllots < ActiveRecord::Migration[5.2]
     create_table :real_allots do |t|
       t.integer :verb_id, null: false
       # 実際の取り組みの配分
-      t.float :allot
+      t.integer :allot
       # 実際の取り組み日数
       t.datetime :term
       t.timestamps
